refactor: 70.J.2 — extract 6 TerminalHandler dispatch groups
Split freminal-terminal-emulator/src/terminal_handler/mod.rs along
dispatch/report/OSC seams. Extracted groups:

  - cursor_ops.rs   — cursor position + visual-style dispatch
  - edit_ops.rs     — erase/insert/delete character dispatch
  - scroll_ops.rs   — scroll region + line-feed + resize dispatch
  - reports.rs      — device attributes, DSR, CPR, color-theme responses
  - window_ops.rs   — CSI t window manipulation dispatch
  - osc.rs          — OSC dispatch (OSC 133/4/10/11/52/...), APC, FTCS

Kept in mod.rs:
  - TerminalHandler struct + full_reset/new/with_scrollback_limit
  - Core data-feeding pipeline (handle_data, placeholder resolution,
    IRM-aware text insertion)
  - set_format + accessors (set_write_tx, take_tmux_reparse_queue,
    current_working_directory, kitty_keyboard_flags)
  - apply_dec_special free function
  - All test modules (moving tests is out of scope for 70.J)

Before: mod.rs = 5,193 lines (single file after 70.I's small edit)
After:  mod.rs = 4,528 lines + 6 focused submodules (~780 lines of
        extracted production code by dispatch responsibility).

No logic or signatures changed — pure move. TerminalHandler fields
changed to pub(super) so sibling modules inside terminal_handler/
can access them without leaking outside the crate.

Verification: cargo test --all green, cargo clippy clean, machete clean.
